refactor: ChangeRequest now bases onto ParcelData, not Parcel
This is in preparation for allowing the reducer to pass in based ChangeRequests into modify methods duing reduction,
so that ChangeRequests can be accessed in modifyUp() etc.

The reducer does not and should not know about Parcel instances, it only deals with ParcelData.
The reason that ChangeRequests even took Parcels rather than ParcelData was to allow them to use the parcel registry
so changes would be based onto the newest known version of the data at that point in the parcel chain.
The only reason there was any doubt that the data was not the newest was due to one of the flaws of the old reducer,
that it did not allow for truly atomic actions, and that pure rendering meant that some parcel could contain stale data.

The registry is no longer being used and will be removed in a future commit.