refactor: remove registry
The parcel registry existed so that when changes were calculated, they would be based onto the newest known version of the data.
The registry contained references to the newest version of each parcel at each point in the parcel chain.
The only reason there was any doubt that the data was not the newest was due to one of the flaws of the old reducer,
that it did not allow for truly atomic actions, and that pure rendering meant that some parcel could contain stale data.
This is no longer a concern with the new reducer.