refactor: actions now use steps instead of keypathmodifiers
KeyPathModifiers were introduced on Actions as a way to bundle the updaters passed from modifyDown() and modifyUp() with the keypaths they were called at,
so that the ChangeRequestReducer could go through and apply each updater at the correct depth.

The reason this is refactored is that KeyPathModifiers had an unforeseen limitation:
 the modifyDown() and modifyUp() updaters did not know their order in comparison to each other.
This has only become a problem now that we want to pass the ChangeRequest into each modifyUp() updater,
and each ChangeRequest needs to know the previous state of the data at exactly the point where its being applied.