refactor: replace ChangeRequest._setBasedParcelData with prevData and nextData
There will be the need to set precalcuated nextData directly on a ChangeRequest. Until now this was not possible,
as a ChangeRequest would be 'based' (given initial data to play actions on top of), and nextData would be calculated from this.
As the _setBasedParcelData request is internal anyway, remove this in favour of a more feature complete _create() method.
_create can accept precalculated nextData. Basing occurs via calling _create({prevData}), and nextData will be calculated and cached as it always has been.