add: ParcelHoc.config.modifyBeforeUpdate
The modifyBeforeUpdate config option allows derived data to be set on a Parcel. Whenever the data in a ParcelHoc is about to be initialised or updated in any way, it is passed through all modifyBeforeUpdate functions.

Each function in the modifyBeforeUpdate array operates just like the updater provided to Parcel.modifyUp().