refactor: add registry back in, remove all references between parcel instances
This is a performance refactor. Previously Parcels contained a reference to the Parcel that created them,
directly and via their onDispatch function, using this to propagate changes back up. This would be fine
except pure rendering in React means that older Parcels inside ParcelBoundaries that never updated still
propagate their changes up via chains of outdated Parcels. For each ParcelBoundary, an entire parcel tree
of oudated references was being kept in memory. The data in this situation was still fine due to the
design of the ChangeRequestReducer

The solution is to reinstate the registry, and only allow Parcels to look up others via that registry.
The only shared references between Parcels are now:
- the registry, which is common and mutated among all Parcels in a tree
- any references pulled into scope by functions passed to Parcel._create, usually _updateChangeRequestOnDispatch
  these are to be kept to an absolute minimum, and NEVER pull in a direct reference to a Parcel