add: ParcelBoundary and ParcelBoundaryHoc onCancel and onRelease
These functions allow the cancel() and release() functions to be hooked into,
to trigger side effects and delay or ignore the act of cancelling or releasing the buffer.
This can really help keep logic in the realm of hocs and data, and out of
the realm of rendering and interaction.