fix: ParcelBoundary now always keeps parcelFromProps updated
ParcelBoundary keeps a reference to the parcel it receives from props
so it can act upon changes to that parcel. Previously this was not being
updated in all cases, so if a keepState enabled ParcelBoundary chose not
to update its inner parcel based on prop changes, then ParcelBoundary would
also not keep record of the fact that the parcel from props had ever changed.

Have not seen any outward effects of this, but logically its a problem that
should be fixed.