break: reveal super edge case that nobody will likely hit
parcels know of each others existence via the registry, which each
parcel adds itself too. this is so even parcels that havend re-rendered
in forever can have a reference to the newest version of their parent
parcel. boundary splitting is making sure that when a parcel boundary
is used, that they each have their own registry, so parel boundaries
dont accidentally share updates telepathically between each other.

right now this is a tad broken with hooks. ill fix it soon, but the
frequency that we'll hit this (i.e. never) means id rather uncover
this logical hole for a bit to get hooks out